Alterando o valor de comparação para decimal
Alterando o valor de comparação para decimal para trabalhar melhor com conversões monetárias, devido a arredondamentos do double.
Ajuste de Endpoints.

+++ CurrencyConverter/CurrencyConverter/Model/Dto/CurrencyToConvertDto.cs
@@ -4,14 +4,14 @@ namespace CurrencyConverter.Model.Dto
     {
         public string From { get; set; }
         public string To { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
         public CurrencyToConvertDto()
         {
             //Empty Constructor
         }
 
-        public CurrencyToConvertDto(string from, string to, double amount)
+        public CurrencyToConvertDto(string from, string to, decimal amount)
         {
             this.From = from;
             this.To = to;
